Make CiviFormProfileData an interface
